refactor(config-ui): remove the version store context
