--- config-ui/src/store/version/context.tsx
@@ -1,46 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-import React, { useContext } from 'react';
-
-import { PageLoading } from '@/components';
-
-import type { VersionType } from './types';
-import { useContextValue } from './use-context-value';
-
-const VersionContext = React.createContext<{
-  version?: VersionType;
-}>({});
-
-interface Props {
-  children?: React.ReactNode;
-}
-
-export const VersionContextProvider = ({ children }: Props) => {
-  const { loading, version } = useContextValue();
-
-  if (loading) {
-    return <PageLoading />;
-  }
-
-  return <VersionContext.Provider value={{ version }}>{children}</VersionContext.Provider>;
-};
-
-export const VersionContextConsumer = VersionContext.Consumer;
-
-export const useVersion = () => useContext(VersionContext);
